Remove neuron daemon start and make Zettelkasten mappings local
The functionality for automatically starting and stopping the neuron
daemon has been migrated to the neuron.vim plugin, so no need for it
to be here anymore.

Zettelkasten-specific gf and gS mappings were interfering with the
original gf and gS in certain edge cases (opening both Zettelkasten and
non-Zettelkasten markdown files). Making the mappings local (i.e. only
valid in the given buffer) seems to have fixed this problem.